+++ lib/web/views/layout_view.ex
@@ -1,0 +1,3 @@
+defmodule Machinery.LayoutView do
+  use Machinery.Web, :view
+end