+++ test/support/test_struct.exs
@@ -1,0 +1,16 @@
+defmodule MachineryTest.TestStruct do
+  import Ecto.Changeset
+  use Ecto.Schema
+
+  schema "test_structs" do
+    field :state, :string
+    field :missing_fields, :boolean
+    field :force_exception, :boolean
+    timestamps()
+  end
+
+  @doc false
+  def changeset(test_struct, attrs) do
+    cast(test_struct, attrs, [:state, :missing_fields, :force_exception])
+  end
+end