+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # ChangeLog
 
 ## WIP Version
-
+- Adding first version of Machinery Dashboard - [Pull Request](https://github.com/joaomdmoura/machinery/pull/14)
 
 ## 0.7.0
 - Improving docs